Queries: reduce number of different tables: value_timestamp is not optional
Reduce the number of possible table configurations by making
value_timestamp mandatory for every datastream table.
The rationale behind this change is to have a simpler codebase at the
cost of some additional storage usage. explicit timestamp is likeley the
most common usecase on devices, so it wouldn't make a really big
difference with the previous approach in terms of storage in real world
scenarios.