message_tracker: pass options to init and use start_link
Previously, AMQPDataConsumer just monitored the MessageTracker and died
in case of :DOWN message, but not the other way around. However, if an
AMQPDataConsumer dies, the delivery tags that the MessageTracker holds
lose meaning since they are valid only in the context of the channel
that was held by AMQPDataConsumer, so it just makes sense that the
MessageTracker is terminated too.  This also simplifies AMQPDataConsumer
exit message trapping since now it just has to worry of channel
monitoring