Add ConsumersSupervisor
The supervision strategy is this: every AMQPDataConsumer is supervised
by AMQPDataConsumer.Supervisor with a one_for_one strategy, so that a
crash in one AMQPDataConsumer does not crash other AMQPDataConsumers.
Then, this supervisor is put in rest_for_one supervision with
ConnectionManager, since if ConnectionManager crashes it is expected
that all AMQPDataConsumer crash because they contain a channel referring
to the connection.  Note that ConnectionManager does not crash when
connection goes down, it just tries to reconnect, but AMQPDataConsumers
will crash nonetheless since their Channels are going to crash. This is
expected and will cause them to wait for the connection to be
established before opening a new channel.