actions: Add Swagger API generation
The docs procedure ignored Swagger API, also because it was pretty much impossible to
handle earlier (maybe?). Anyway: this ensures that the api/ directory is always in sync
with a specific swagger-ui version, and also ensures that we have the swagger UI index
file safely stored in our source repository.

Fixes #337

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>