appengine_api: always consider push to properties interfaces succesful
A push can fail only if the device doesn't have a session on the broker. But if
it this happen, then the device will ask for all properties at the next
reconnection. This allows to push properties to devices before their first
connection, making sure they're saved to the database.

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>