data_updater_plant: declare custom AMQP exchanges on trigger load
When an AMQP trigger with a custom exchange is loaded, the Data Updater process
will now ask TriggersHandler to register its target, which in turn will declare
the exchange through AMQPEventsProducer. This needs AMQPEventsProducer to be
started before the consumers, so we swap them in the supervision tree.
For now, the exchanges remain there forever. We need to decide who will be
responsible to clear them once they're not needed anymore.

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>