data_updater_plant: fix queue block during ask_clean_session
The return value of ask_clean_session was ignored before, so it did not crash
the process. 5beed84913138bb5ecaa902b1bb64435b191df14 introduced the use of the
returned value, so the process now crashes if an error is returned, which was
happening when we tried to disconnect a device that was already disconnected.
This fixes that problem, making sure to return :ok in that case.

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>