docker-compose: avoid accidentally creating housekeeping_public.pem
The short syntax for docker-compose volumes creates the source on the
host if it does not exist. This meant that user that accidentally tried
to launch docker-compose up before executing docker-compose-initializer
found themselves with a housekeeping_public.pem directory, that was seen
as an existing file by the initializer and left the setup in an invalid
state.
Using the long syntax avoids the silent creation of the directory and
just makes the docker-compose up command fail (so the user can read the
instructions again and find docker-compose-initializer).
The long volume syntax requires at least compose syntax v2.3, so we also
have to bump that.

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>