device: don't discard opts when retrieving object aggregations
The call to retrieve_endpoint_values was discarding opts when multiple paths
were present in an object aggregation (e.g. when it contained multiple instances
of a parametric path). This caused the call to always have false in the
explicit_timestamp option, which in turn caused a crash when an interface with
explicit_timestamp true was queried due to different timestamp columns.

Fix #480

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>