appengine: allow returning nil in AstarteValue.to_json_friendly
The interaction between https://github.com/astarte-platform/astarte/issues/550
and https://github.com/astarte-platform/astarte/pull/551 was preventing
aggregate interfaces that got populated with incomplete objects to be queried.
This fixes that, allowing AppEngine to return null in some JSON fields (since
it's not its job to fix that kind of issue anyway).

Signed-off-by: Riccardo Binetti <riccardo.binetti@ispirata.com>