Trigger Engine: add support for trigger policy
Add support for customizable redelivery policies for HTTP triggers.
TriggerEngine retries (or not) to deliver a payload according to the
related trigger policy. Default to not retrying on any error, as not to break
compatibility.

There is currently only one listener on the exchange where DUP publishes data.
In the future, there will be one listener per policy queue.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>