TriggerEngine: let policy process crash on unexpected error
Policy process did catch unexpected message delivery errors, but had no info
on whether to redeliver the message or not. Let the policy process crash
instead of handling the unexpected error, and let the related AMQPConsumer
process come down with it. This will put handled but unacked messages (if any)
back in the queue.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>