Realm Management: introduce experimental trigger policy prefetch count
When using Trigger Delivery Policies, ordered event delivery is not guaranteed
because messages might not be dequeued (thus processed) sequentially.
As a possible solution, allow the policy to specify the AMQP consumer prefetch
count used for its event queue. If the value is 1, an instance of Trigger Engine
processes events sequentially.
This is an experimental feature and thus disabled by default. You can enable it
with the `REALM_MANAGEMENT_ALLOW_TRIGGER_POLICY_PREFETCH_COUNT` env var.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>