RM: Add device deletion logic
Device deletion begins when a device_id is put in the
`deletion_in_progress` table.
The actual deletion of data is started only when it has been
acknowledged by both DUP (two times) and the broker (once).
The DeviceRemoverSupervisor supervises DeviceRemover tasks.
Since deletion is asynchronous, it may happen that the
RM service goes down during deletion, or that a temporary filure
brings down the deletion service. Therefore, a supervision tree
is put in place.
Data deletion is started (or resumed, if needed) by the
DeviceRemoval.Scheduler and it is carried out by the
DeviceRemoval.DeviceRemover Task.
The DeviceRemoverSupervisor supervises DeviceRemovers.