HK: support device_registration_limit update
Allow to use the update realm PATCH REST API to
update the device registration limit of a realm.
If the limit is set to null, the exisiting value
will be removed. If it is set to a non negative
integer, its value will be the new limit.

No checks are performed on the number of devices
already registered in the realm.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>