DUP: do not crash when purge properties payload is not zlib deflated
Due to a quirk in OTP's zlib library, the safe_inflate function
can actually throw an error if the payload is malformed.
The unhandled error made the DataUpdater process crash, resulting
in the malformed purge properties message being requeued and
so on and so forth.
Fix this behaviour by wrapping :zlib calls in a try-catch block.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>