Consider the known_value field of triggers when relevant
The field was wrongly discarded by the validation logic since the latter
expects the interface definition as an additional argument when
validating trigger fields that relate to the interface; when the
interface definition is not provided, the validation defaulted to
discarding the known_value field.
The issue is fixed by not discarding the field in such cases where the
exact field type cannot be retrieved.

Closes #393

Signed-off-by: Davide Briani <davide.briani@secomind.com>