Update eslint configuration and remove react-scripts dependency
Update the plugins and configurations for eslint to function without the
dependencies included in react-scripts, then remove the react-scripts
dependency.
Specifically, the set of rules such as the `airbnb` code style are
removed in favor of modern default sets.

Changed with:
```
npm install eslint-config-react-app eslint-plugin-cypress \
eslint-plugin-import eslint-plugin-testing-library

npm uninstall eslint-config-airbnb @typescript-eslint/eslint-plugin \
@typescript-eslint/parser eslint-import-resolver-typescript \
eslint-plugin-import eslint-plugin-jsx-a11y eslint-plugin-react \
react-scripts
```

The dependencies needed to build the project are moved from the
devDependencies list to the dependencies list.
TypeScript is also updated to the latest version to properly support the
changed dependencies.

Changed with:
```
npm install typescript@5.3.3
```

Signed-off-by: Davide Briani <davide.briani@secomind.com>