Update React JSON Form and AJV dependencies
Updated with:
```
npm install @rjsf/bootstrap-4@5.17.0 @rjsf/core@5.17.0 ajv-draft-04 \
ajv@8.12.0 @rjsf/validator-ajv6
```

The code is updated to keep support for JSON Schema draft-04, which now
requires additional specific dependencies.
JSON Schema draft-04 is a dated version but it's the one still used by
the definitions of Astarte Blocks in Astarte Flow, so it's the supported
version for the moment.

Signed-off-by: Davide Briani <davide.briani@secomind.com>