RM: refute interfaces with storage TTL > max TTL
There were no checks on RM side on whether an interface
contained only mappings with database_retention_ttl lower
than the maximum storage TTL allowed for the realm, making it
possible to bypass the maximum allowed retention. Add the
check on interface install and update.

The code could be better (e.g. we return :ok/{:error, reason}
where {:ok, true/false}/{:error, reason} would be more appropriate),
due to the fact that it has to interact with legacy sections.

This change can be breaking, as some interfaces won't be installed
anymore if the max allowed retention for the realm is set.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>