Workflows: allow to publish images to Dockerhub again
Workflows publishing images to Dockerhub did not run up to the
publishing step because they depended on the apps test workflow,
which did always fail because it could not access the
codecov token secret when called via another workflow.
Allow the apps test workflow to inherit secrets from the calling
workflow, so that it can successfully run again and unlock
the publisher workflow.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>