refactor(appengine): port fetch_public_key to exandra
this function needs some special attention because by default Ecto SQL
raises if we get an error back from exandra.

as this function is also supposed to handle the situation where the
keyspace does not exist, which results in a Xandra.Error being generated,
we need to use a _workaround_ which lets us handle the error ourselves.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>