Migrate unregister_device queries to Ecto
Use Ecto to compose and run the query.

This introduces a change in how the API responds to user requests for
unregistering devices.

Before this change, unregistering an existing device returned
successfully regardless of the device registration status,
and unregistering a non-existing device returned `{:error,
:device_not_registered}` which was translated as a 403 Forbidden error
from Pairing API.

Now, the idempotency on existing devices is kept as before, but Pairing
returns `{:error, :device_not_found}` for a non-existing device, which
is translated by Pairing API as a 404 Not Found error.

This change better reflects the error that actually happened instead of
giving misleading feedback to API users. Moreover, for all concerns
about authorization or security concerns, Pairing API already returns
403 Forbidden errors when the API user requests an operation on a device
that is not allowed with the permissions contained in the auth token.

This commit also ensures that `{:error, :device_not_found}` errors are
correctly handled in a similar way by Pairing API in the case of other
operations that return that error. Indeed, the APIs to get device
credentials, verify device credentials, and get device info, may all
return such an error, as the change was introduced in PR #1082.

Signed-off-by: Davide Briani <davide.briani@secomind.com>