DUP tests: consistently define object aggregated value types
The `/value` endpoint in the `com.example.TestObject` interface
is defined to handle `double` data. However, in tests it received
an integer value. This worked due to a bug in the CQex driver.

Be conistent with the type declared and use a float instead of
an integer, and update the bytes count consequently.

Similar to https://github.com/astarte-platform/astarte/pull/1044.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>