refactor(appengine): centralize insert to sql definition
using the standard ecto functions, there is no way to get the sql and
params for an insertion from the adapter, which is instead the case for
every other query type.

this pr defines a function, `insert_to_sql`, which behaves a bit like
`to_sql` for insertions.

instead of using `Exandra.Connection.insert`, the logic was re-implemented
here as for the little effort of copy-pasting the logic here, we
get the added benefit of being able to declare custom syntax for the
parameters, like using `intAsBlob(?)` for an insertion in the KvStore,
while exandra is limited to just `?` directly.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>