Avoid catching Xandra errors, let RPC take care of temporary failures
This change avoids catching and managing Xandra errors within the code
dedicated to interact with the database, where the concern does not
require handling expections.

Exceptions and temporary failures are better handled by whomever calls
Pairing functionality: in this case RPC, in the future it might be
directly a Phoenix conn handling a request.

Note that Astarte RPC already handles failures by responding with a
generic error to the caller.
The exception is logged for operational purposes, so developers can
troubleshoot the issue, but these kind of database errors are not of
much use to API users.

Signed-off-by: Davide Briani <davide.briani@secomind.com>