fix(dup): don't crash on register_device_with_interface
previously string keys were given to exandra in `insert_all`.
however, when given an actual schema (ie, not a string with a table name),
ecto expects atom keys for the values.

this resulted in a crash with the log
%ArgumentError{message: \"unknown field `\\\"group\\\"` in schema
Astarte.DataUpdaterPlant.DataUpdater.KvStore given to insert_all. ...}

the fix is to just use atom keys

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>