test(te): per-module astarte instance using mocks
assign a random `astarte_instance_id` to each test module so that they
can be run in parallel without interference.

this is achieved by mocking the Config so that it's astarte_instance_id
functions return the custom value.

all database_test_helper definitions had to be changed from static
strings to functions so that the value of `Realm.keyspace_name` depends
on the mocked value of Astarte.DataAccess.Config for the current process.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>