feat(AE): libcluster
Adds `libcluster` in `AppEngine` to allow for node discovery.

The library is used to discover other nodes in the cluster at runtime, currently 2 strategies are introduced:

- none :: with this strategy nodes of the cluster do not try to find
  each other, keeping a retrocompatible approach to clustering.
- kubernetes :: with this strategies DUP nodes will try to find other
  instances of DUP view kubernetes API (querying the endpoints endpoint).

This should not be a breaking change as the default clustering strategy
is still to use no clustering at all (CLUSTERING_STRATEGY environment
variable defaults to none).

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>