feat(DUP): dynamic `DataUpdater`s
Moves DataUpdater process to a Horde dynamic supervisor (distributed)
and a Horde registry (also distributed). This allows to abstract the
supervision tree of DataUpdater processes away from a single instance of
DUP (when clusterized) and to distribute the workload of handling
messages across different replicas of DUP, indipendently from there the
AMQPDataConsumer actually is.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>