Add clustering strategy for AppEngine in Docker Compose
Define a new clustering strategy that works in Docker Compose. The
strategy assumes a specific service name, and relies on DNS queries to
resolve the service IP address dynamically.
The strategy is used by Appengine to find and connect to Data Updater
Plant instances in order to perform RPC operations.

Signed-off-by: Davide Briani <davide.briani@secomind.com>