feat(policy): init
this adds a generator for policies.

to allow for an efficient generation of handlers, the previous
`ErrorRange` and `ErrorKeyword` generators were not used.

This is because of the requirement that there must be no intersection
between the handled range for each handler, and the fact that there is a
small range of possible errors

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>