fix(ae): properly declare needed exchanges
AMQPClient needs an exchange `events_exchange_name` to be declared,
otherwise the queue binding fails and the whole application fails to
start.

this was mitigated in CI with a custom exchange declaration and in
production by having DUP or TE declare this exchange for AE before it
could start successfully.

while for queues it's documented
(eg [here](https://www.rabbitmq.com/tutorials/amqp-concepts#queues))
that "The declaration will have no effect if the queue does already
exist and its attributes are the same as those in the declaration", this
appears to be never explicitly stated to also be the case for exchanges.

however, in astarte we're already declaring this exchange twice:
- once in [trigger engine](https://github.com/astarte-platform/astarte/blob/d5ed3c0a6c1af3b620a436ae506462c756537f6d/apps/astarte_trigger_engine/lib/astarte_trigger_engine/amqp_consumer/amqp_message_consumer.ex#L171)
- once in [data updater plant](https://github.com/astarte-platform/astarte/blob/d5ed3c0a6c1af3b620a436ae506462c756537f6d/apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/amqp_events_producer.ex#L120)

if it has not caused us problems thus far, it should also work as
expected declaring it a third time

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>