refactor(housekeeping)!: move realm update validation to the api
previously the update_changeset was treated as a normal changeset,
casting all values and validating parameters which are not even valid
update parameters

the realm would then be sent to the backend which would fail if
the parameters were actually changed

with the current rework, the api only allows updating of the valid
parameters

fixes the situation in which the changes included a realm_name, in
which case the update would refer to the realm in params.data.realm_name
instead of params.realm_name

BREAKING CHANGE: previously it was possible to specify the realm to
update by setting `data.realm_name`, as it would be applied by the
changeset and passed on to the backend. Now, the update can only
refer to the realm from the path.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>