feat(rm): remove delete_trigger rpc
the internal state of the realm_management trigger has been updated to
also include trigger_uuid and simple_triggers_uuids from the core struct.

this was needed for `delete_trigger` to avoid reading the trigger
from the database twice.

it should not be visible to the user as the json api only exposes a
select few fields from the struct anyway.

install_trigger was updated to return the updated trigger information
from the db insertion

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>