Pairing API: Add API route plug to verify realm exists
Before running the business logic of API routes that involve queries
against realm resources, use a Plug to verify whether the realm
specified in the API route indeed exists.
If the realm does not exist, the request is terminated right away with
a 403 HTTP error.

This approach allows to leave the low-level code that builds and runs
database queries simple and without employing defensive programming to
handle the fact that the realm/keyspace may not exist: at that level the
realm/keyspace is expected to exist and the database query rightly
raises in case of error.
The API user is still provided with a nicely formatted error response.

Existing tests are updated or expanded to reflect the expected
behaviors: low level queries raise in case of non-existing
realm/keyspace, but HTTP API requests receive well-formatted error
messages indicating a client error (4xx) instead of a server error
(5xx).

Signed-off-by: Davide Briani <davide.briani@secomind.com>