refactor(housekeeping): remove xandra queries (#1406)
* refactor(HK): Function for `insert_datastream_max_retention`

Use KvStore.insert for setting datastream maximum storage retention

Replace direct Xandra query with KvStore.insert/2 for consistency
with other KV store operations in the codebase.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

* refactor(HK): Function for `insert_realm`

Update the `insert_realm` function to use Ecto queries instead of
Xandra queries.

The goal of this is to remove the direct use of Xandra queries in the
and to use Ecto queries instead.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

* refactor(HK): Function for `insert_realm_astarte_schema_version`

Use KvStore.insert for realm schema version insertion

Replace direct Xandra query with KvStore.insert/2 for consistency
with other KV store operations in the codebase.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

* refactor(HK): Function for `insert_realm_public_key`

Use KvStore.insert for realm insert public key

Replace direct Xandra query with KvStore.insert/2 for consistency
with other KV store operations in the codebase.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

* refactor(HK): Function for `create_astarte_kv_store`

Update the `create_astarte_kv_store` function to use Ecto queries
instead of Xandra queries.

The goal of this is to remove the direct use of Xandra queries in the
and to use Ecto queries instead.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

* refactor(HK): Function for `create_realms_table`

Update the `create_realms_table` function to use Ecto queries
instead of Xandra queries.

The goal of this is to remove the direct use of Xandra queries in the
and to use Ecto queries instead.

Signed-off-by: Omar <omar.brbutovic@secomind.com>

---------

Signed-off-by: Omar <omar.brbutovic@secomind.com>