refactor(HK): Function for `set_schema_version`
Update `migrate_astarte_keyspace_from_version` and
`migrate_realm_from_version` to use `KvStore.insert`
instead of direct Xandra calls.
Remove the need for `use_keyspace` and pass the keyspace_name
as an argument to the functions.

The goal of this is to remove the direct use of Xandra queries in the
and to use Ecto queries instead.

Signed-off-by: Omar <omar.brbutovic@secomind.com>