fix(dup): crash when performing updates of deleted values
previously, updates were changed from upserts to normal updates to
avoid re-inserting previously deleted values. This introduced a bug
because Ecto would complain about stale values if the update did not
modify the struct in the database. fix this by setting the appropriate
flag (`allow_stale`)

regression tests for previous fixes have been added to prevent this
from happening again

the update functions now return `:ok`. this does not change the usage
of the functions as the return value was just discarded, but returning
the deletion in progress entry when the value did not exist in the
database is incorrect

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>