feat: Add env var for vernemq kubernetes service name
Previously the Kubernetes service name used to look for VerneMQ
instances was hardcoded to the fixed value `astarte-vernemq`.
However, `astarte` is just the default prefix and VerneMQ replicas are
not found in Astarte instances that have a name other than `astarte`.
This change introduces a new environment variable,
`VERNEMQ_CLUSTERING_KUBERNETES_SERVICE_NAME`, that can be set to specify
a custom Kubernetes service name, if needed.

Signed-off-by: Davide Briani <davide.briani@secomind.com>