fix(dup): ensure correct bson types for properties on device connection
previously, we would just send the values as they were gathered from
the database. this works fine, but this historically had issues with
a couple value types:
- binaries are not wrapped in a `%Cyanide.Binary{}`, so they're sent
  as strings
- datetimes are read from the database as integers, so they're sent as
  timestamps.

the issue with datetimes was resolved by using exandra instead of
cqex

to solve the issue with binaries, we ensure `send_value` uses the
correct bson type.

to test this, a server owned property interface is added with a mapping
of each value type

fixes #935

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>