fix(rm): faulty checks for existing database resources (#1494)
There were some faulty checks for existing database resources, due
to the used `Repo.some?` custom function which now returns a `{:ok,
bool}` tuple instead of directly a boolean, as in previous iterations.

Indeed, the tuple was either checked in `if` statements where it always
computed to a trueish value, of it was pattern matched to boolean values
which resulted in an match error and a crash.

The latter one is also the case of a bug in the DeviceRemover task,
which crashed just after launch and resulted in devices to be
stuck in the "in deletion" status.

Fixes #1493.

Signed-off-by: Davide Briani <davide.briani@secomind.com>