docs: add description for ASTARTE_INSTANCE_ID variable
This patch introduces documentation for the optional ASTARTE_INSTANCE_ID
variable. The variable defaults to an empty string for backward
compatibility, but can be set to improve namespace isolation when running
multiple Astarte instances on a shared database. The Operator fully
supports this variable and ensures that Cassandra keyspace names remain
within valid length limits.

Signed-off-by: Eddy Babetto <eddy.babetto@secomind.com>