fix(housekeeping): migration task (#1515)
when CSystem calls were updated to use ecto, we lost track of the
`use_keyspace` function which made migrations work with xandra cluster.

update migrations to include a `:keyspace` placeholder which is
substituted with the keyspace name by the migrator, as tests with the
`prefix` option for csystem were unsuccessful.

split 0006 into two migrations as only one query can be put into each
migratrion

add an actual test for the migratior which runs migrations, and remove
a test which did nothing (destroyed a table in the astarte keyspace and
tried to run realm migrations)

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>