feat: device deletion finished triggers (#1546)
the wrong event type was used when encoding events, so it would result
in a failure when trying to encode the event

also, the decoded device id was used to dispatch the event, with the
library expecting the hardware id

we do this in dup after the device has been disconnected, so that we
make sure the group information cannot change after replicating the
information.

test files have been modified a little to allow for cleaner tests and
allowing deletion tests to run on their own

there is only one instance which made the process fail if the device
was no longer in the database, and it was when retrieving the device
introspection.

Returning an empty map when the device can't be found removes the
special handling for this situation


Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>