feat(events): handle volatile triggers (#1558)
as the volatile triggers are not fetched from the database, the cache
can be optimized by including the event_key in the cache key.

the cache structure is optimized for reading data, as that is by far the
most used operation on the trigger cache.

an additional cache is needed for the deletion operation (in order to
avoid having to iterate the whole ets table to find the trigger_id).

volatile triggers don't use a tll as they can't be read from the database
at expiry.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>