refactor(housekeeping): return 409 for existing realms
we're already returning 409 in realm management for existing interfaces
and triggers, so this would be closer in spirit to what we're already
doing in other services

somewhere between 1.2 and 1.3, the error returned went from
"unprocessable entity" to "service unavailable".

"unprocessable entity" makes no sense because the input format is
correct, so instead of switching one wrong error for another, we can
use a proper error instead.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>