fix(dup): fix possible crash on interface-specific volatile triggers
previously, when
1. installing a volatile trigger for a specific interface
2. the interface results in a cache miss but is valid for the device

the volatile trigger would be installed but the device status would not
be updated.

this results in a crash when the trigger generated events, as the logic
expects the interfaces to be in the data updater state.

by just loading the interfaces normally, we avoid this possible crash

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>