refactor(pairing): rework ownership voucher api
now the `/ownership` endpoint both saves the voucher to database and
performs to0 with the rendezvous.

the voucher data saved to the database is now cbor encoded, respecting
the "blob" database type.

a new module is initialized for the request parameters, just like
device registration request and credentials request, which allows
using ecto changesets for the data validation.

the expected format for the given ownership voucher MUST now contain
the "BEGIN OWNERSHIP VOUCHER" and "END OWNERSHIP VOUCHER" lines,
and only text in between said lines is used for the ownership voucher.

tests have been removed from the ownership voucher file as the module
is no longer the one performing data validation.

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>