chore: rudimentary session management (#1607)
create a session as defined by section 4.3 of the FDO 1.1 spec using
a single identifier in the authorization header

the current implementation works but has a few caveats:
- it stores the private key to the database (we never did it!)
- uses an uuid4 as the identifier, which might not be the best solution

alternatives considered: storing the sessions in cache
unfortunately, we need the database to store sessions because we don't
have any guarantee the same pairing replica will handle subsequent
messages.

an horde registry might work tho!

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>