fix: crashes on new scylladb versions (#1711)
somewhere along the lines scylladb changed the default for 
NetworkReplicationStrategy to use tablets. these are incompatible with
some astarte queries as we need lightweight transactions for some
operations (such as device deletion)

for now, always disable tablets so that the old logic still works.
this is accomplished by always trying to create keyspaces and explicitly
disabling tablets, and in case we're on cassandra or an old version of
scylladb, we catch the failure and try again without specifying an
option (as tablets don't exist at all in those databases)

update scylladb to the new, source available version to align it to the
production environment

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>